circuit-types: v2: Add v2 balance and deposit
