+++ circuit-types/src/v2/mod.rs
@@ -1,1 +1,17 @@
 //! V2 circuit types
+
+pub mod balance;
+pub mod csprng_state;
+pub mod deposit;
+pub mod state_wrapper;
+
+use constants::Scalar;
+
+// ----------------
+// | Type Aliases |
+// ----------------
+
+/// A commitment to a state element
+pub type Commitment = Scalar;
+/// A nullifier for a state element
+pub type Nullifier = Scalar;
